azure/key_vault/requirements.txt
python-dotenv
azure-identity
azure-keyvault-secrets